+++ src/input/inputFileParser/QMInputParser.cpp
@@ -90,6 +90,12 @@ QMInputParser::QMInputParser(Engine &engine) : InputFileParser(engine)
     );
 
     addKeyword(
+        std::string("mace_model_path"),
+        bind_front(&QMInputParser::parseMaceModelPath, this),
+        false
+    );
+
+    addKeyword(
         std::string("slakos"),
         bind_front(&QMInputParser::parseSlakosType, this),
         false
@@ -309,14 +315,32 @@ void QMInputParser::parseMaceModelSize(
     else if ("medium_omat_0" == size)
         QMSettings::setMaceModelSize(MEDIUMOMAT0);
 
+    else if ("custom" == size)
+        QMSettings::setMaceModelSize(CUSTOM);
+
     else
         throw InputFileException(std::format(
             "Invalid mace_model_size \"{}\" in input file.\n"
             "Possible values are: small, medium, large, small-0b,\n"
             "medium-0b, small-0b2, medium-0b2, large-0b2, medium-0b3,\n"
-            "medium-mpa-0, medium-omat-0",
+            "medium-mpa-0, medium-omat-0, custom",
             lineElements[2]
         ));
+}
+
+/**
+ * @brief parse external MACE model url
+ *
+ * @param lineElements
+ * @param lineNumber
+ */
+void QMInputParser::parseMaceModelPath(
+    const std::vector<std::string> &lineElements,
+    const size_t                    lineNumber
+)
+{
+    checkCommand(lineElements, lineNumber);
+    QMSettings::setMaceModelPath(lineElements[2]);
 }
 
 /**